--- core/src/plugins/index.ts
@@ -1,2 +1,0 @@
-export { default as SSRWebpackRunPlugin } from './webpackRun';
-export { default as SSRWebpackPlugin } from './plugin';
