+++ packages/kkt-plugin-ssr/package.json
@@ -20,11 +20,12 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "webpack": "5.69.1",
-    "@types/webpack-node-externals": "^2.5.3"
+    "kkt": "7.1.5",
+    "@types/webpack-node-externals": "2.5.3"
   },
   "dependencies": {
-    "webpack-node-externals": "3.0.0",
-    "webpack-merge": "5.8.0"
+    "@kkt/simple-progress-webpack-plugin": "1.1.5",
+    "webpack-merge": "5.8.0",
+    "webpack-node-externals": "3.0.0"
   }
 }
